combine svelte example, add custom component
