+++ examples/svelte/sveltekit/src/lib/Builder.svelte
@@ -1,0 +1,64 @@
+<script context="module" lang="ts">
+	export const prerender = true;
+</script>
+
+<script lang="ts">
+	import { page } from '$app/stores';
+
+	import Counter from '$lib/Counter.svelte';
+
+	import * as BuilderSDK from '@builder.io/sdk-svelte';
+
+	const CUSTOM_COMPONENTS = [
+		{
+			component: Counter,
+			info: {
+				name: 'Counter',
+				inputs: [
+					{
+						name: 'name',
+						type: 'string',
+						defaultValue: 'hello'
+					},
+					{
+						name: 'count',
+						type: 'number',
+						defaultValue: 0
+					}
+				]
+			}
+		}
+	];
+	// TODO: enter your public API key
+	const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
+
+	let content = undefined;
+	let canShowContent = false;
+	const fetch = async () => {
+		content = await BuilderSDK.getContent({
+			model: 'page',
+			apiKey: BUILDER_PUBLIC_API_KEY,
+			options: BuilderSDK.getBuilderSearchParams(
+				BuilderSDK.convertSearchParamsToQueryObject($page.url.searchParams)
+			),
+			userAttributes: {
+				urlPath: $page.url.pathname
+			}
+		});
+		canShowContent = content || BuilderSDK.isEditing();
+	};
+
+	fetch();
+</script>
+
+<div>Hello world from your SvelteKit project. Below is Builder Content:</div>
+
+{#if canShowContent}
+	<div>page: {(content && content.data && content.data.title) || 'Unpublished'}</div>
+	<BuilderSDK.RenderContent
+		model="page"
+		{content}
+		apiKey={BUILDER_PUBLIC_API_KEY}
+		customComponents={CUSTOM_COMPONENTS}
+	/>
+{/if}
