+++ examples/svelte/sveltekit/src/routes/[slug].svelte
@@ -3,30 +3,7 @@
 </script>
 
 <script lang="ts">
-	import { page } from '$app/stores';
-
-	import * as BuilderSDK from '@builder.io/sdk-svelte';
-
-	// TODO: enter your public API key
-	const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
-
-	let content = undefined;
-	let canShowContent = false;
-	const fetch = async () => {
-		content = await BuilderSDK.getContent({
-			model: 'page',
-			apiKey: BUILDER_PUBLIC_API_KEY,
-			options: BuilderSDK.getBuilderSearchParams(
-				BuilderSDK.convertSearchParamsToQueryObject($page.url.searchParams)
-			),
-			userAttributes: {
-				urlPath: $page.url.pathname
-			}
-		});
-		canShowContent = content || BuilderSDK.isEditing();
-	};
-
-	fetch();
+	import Builder from '$lib/Builder.svelte';
 </script>
 
 <svelte:head>
@@ -45,12 +22,7 @@
 		to your new<br />SvelteKit app
 	</h1>
 
-	<div>Hello world from your SvelteKit project. Below is Builder Content:</div>
-
-	{#if canShowContent}
-		<div>page: {(content && content.data && content.data.title) || 'Unpublished'}</div>
-		<BuilderSDK.RenderContent model="page" {content} api-key={BUILDER_PUBLIC_API_KEY} />
-	{/if}
+	<Builder />
 </section>
 
 <style>
